refactor(examples): update import path for timer555 asset
The asset was moved to a local directory, so the import path needs to be updated to reflect the new location.
