+++ src/examples/assets/timer-555.ts
